fix: merge AddContact into AddContactButton
