--- src/components/AddContact/index.ts
@@ -1,1 +1,0 @@
-export { AddContact } from './AddContact'
